chore(atlauncher-scripts): add in license header from react-scripts

+++ packages/atlauncher-scripts/bin/atlauncher-scripts.js
@@ -1,46 +1,63 @@
 #!/usr/bin/env node
 
+/**
+ * Copyright (c) 2015-present, Facebook, Inc.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ */
+
+const chalk = require('chalk');
+const utils = require('../utils');
 const spawn = require('cross-spawn');
 
-const script = process.argv[2];
-const args = process.argv.slice(3);
+const args = process.argv.slice(2);
 
-switch (script) {
-    case 'lint':
-    case 'test':
-    case 'build':
-    case 'clean':
-    case 'babel-run': {
-        const result = spawn.sync('node', [require.resolve(`../scripts/${script}`)].concat(args), {
-            stdio: 'inherit',
-        });
+if (!args.length) {
+    console.log(chalk`{white.bgRed No arguments passed in.}`);
 
-        if (result.signal) {
-            if (result.signal === 'SIGKILL') {
-                console.log(
+    process.exit(1);
+}
+
+const scripts = utils.getScripts();
+const scriptIndex = args.findIndex((arg) => scripts.includes(arg));
+const script = scriptIndex === -1 ? args[0] : args[scriptIndex];
+const nodeArgs = scriptIndex > 0 ? args.slice(0, scriptIndex) : [];
+
+if (scripts.includes(script)) {
+    const result = spawn.sync(
+        'node',
+        nodeArgs
+            .concat(require.resolve(`../scripts/${script}`))
+            .concat(args.slice(scriptIndex + 1)),
+        { stdio: 'inherit' }
+    );
+
+    if (result.signal) {
+        if (result.signal === 'SIGKILL') {
+            console.log(
+                chalk.white.bgRed(
                     'The build failed because the process exited too early. ' +
-                        'This probably means the system ran out of memory or someone called ' +
-                        '`kill -9` on the process.'
-                );
-            } else if (result.signal === 'SIGTERM') {
-                console.log(
+                    'This probably means the system ran out of memory or someone called ' +
+                    '`kill -9` on the process.'
+                )
+            );
+        } else if (result.signal === 'SIGTERM') {
+            console.log(
+                chalk.white.bgRed(
                     'The build failed because the process exited too early. ' +
-                        'Someone might have called `kill` or `killall`, or the system could ' +
-                        'be shutting down.'
-                );
-            }
-            process.exit(1);
+                    'Someone might have called `kill` or `killall`, or the system could ' +
+                    'be shutting down.'
+                )
+            );
         }
 
-        process.exit(result.status);
-
-        break;
+        process.exit(1);
     }
-    default: {
-        console.log(`Unknown script "${script}".`);
-        console.log('Perhaps you need to update @atlauncher/atlauncher-scripts?');
-        console.log('See: https://github.com/ATLauncher/style-guide');
 
-        break;
-    }
+    process.exit(result.status);
+} else {
+    console.log(chalk`{white.bgRed Unknown script "${script}".}`);
+
+    process.exit(1);
 }
